#96 move GraphicalDslInstanceShortInfo to format package
