+++ api/common/src/main/scala/de/htwg/zeta/common/format/GraphicalDslInstanceShortInfo.scala
@@ -1,4 +1,4 @@
-package de.htwg.zeta.common.models.project.instance
+package de.htwg.zeta.common.format
 
 import java.util.UUID
 
@@ -12,9 +12,9 @@ import play.api.libs.json.__
 /**
  * Represents concise information on a model, used for REST-API overview uri
  *
- * @param id          the id of the model
- * @param metaModelId the name of the MetaModel
- * @param name        the name of the Model
+ * @param id          the id of the graphical dsl instance
+ * @param metaModelId the id of the graphical dsl
+ * @param name        the name of the graphical dsl instance
  */
 case class GraphicalDslInstanceShortInfo(id: UUID, metaModelId: UUID, name: String)
