+++ api/server/app/de/htwg/zeta/server/controller/webpage/WebpageController.scala
@@ -8,8 +8,8 @@ import scala.concurrent.Future
 
 import com.mohiva.play.silhouette.api.actions.SecuredRequest
 import controllers.routes
+import de.htwg.zeta.common.format.GraphicalDslInstanceShortInfo
 import de.htwg.zeta.common.format.ProjectShortInfo
-import de.htwg.zeta.common.models.project.instance.GraphicalDslInstanceShortInfo
 import de.htwg.zeta.persistence.accessRestricted.AccessRestrictedGraphicalDslRepository
 import de.htwg.zeta.persistence.accessRestricted.AccessRestrictedGraphicalDslInstanceRepository
 import de.htwg.zeta.server.silhouette.ZetaEnv
