+++ api/server/app/views/webpage/WebpageDiagramsOverview.scala.html
@@ -1,8 +1,7 @@
+@import de.htwg.zeta.common.format.GraphicalDslInstanceShortInfo
 @import de.htwg.zeta.common.format.ProjectShortInfo
 @import de.htwg.zeta.common.models.entity.User
 @import de.htwg.zeta.common.models.project.GdslProject
-@import de.htwg.zeta.common.models.project.instance.GraphicalDslInstanceShortInfo
-
 @(user: Option[User], metaModels: Seq[ProjectShortInfo], metaModel: Option[GdslProject], modelInstances: Seq[GraphicalDslInstanceShortInfo])
 
 @styles = {
